requirements.txt
Django==2.2.3
django-crispy-forms==1.7.2
django-filter==2.1.0
Pillow==6.0.0
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
